Calculate the MiniMax Statistics

Given a data frame of pathway-level p-values, calculate the MiniMax statistic (but not the significance level of this statistic).

MiniMax_calculateStatistic(res_df, orderStat = 2L)

Arguments

res_df
A data frame of p-values. The rows correspond to gene sets / pathways and the columns correspond to a data platform for the disease of interest.

orderStat
How many platforms should show a biological signal for a pathway / gene set to have multi-omic "enrichment"? Defaults to 2. See "Details" for more information.

Value

A vector of the MiniMax statistic values.

Details

The MiniMax statistic is defined as the minimum of all pairwise maxima of pathway p-values. This operation is arithmetically equivalent to sorting the p-values and taking the second smallest. In our experience, setting this "order statistic" cutoff to 2 is appropriate for =< 5 data platforms. Biologically, this is equivalent to saying "if this pathway is dysregulated in at least two data types for this disease / condition, it is worthy of additional consideration". In situations where more than 5 data platforms are available for the disease of interest, we recommend increasing the orderStat value to 3.

Examples

  MiniMax_calculateStatistic(multiOmicsHighSignalResults_df[, -(1:2)])

#>  [1] 0.757 0.701 0.697 0.000 0.000 0.687 0.688 0.326 0.275 0.423 0.240 0.706
#> [13] 0.548 0.561 0.538 0.000 0.475 0.684 0.000 0.370 0.586 0.725 0.478 0.031
#> [25] 0.647 0.101 0.391 0.756 0.482 0.509 0.468 0.027 0.725 0.518 0.225 0.524
#> [37] 0.269 0.316 0.720 0.117 0.553 0.411 0.314 0.879 0.210 0.519 0.766 0.618
#> [49] 0.152 0.000