Example Low-Signal Multi-Omics Pathway Results

An example of the input data to the MiniMax statistic with weak biological (synthetic) signal

multiOmicsLowSignalResults_df

Format

A tibble (a modern data frame created by the tibble package) with five columns:

terms the name of a pathway / gene set; this column is called 'terms' for compatibility with pathwayCollection objects from package pathwayPCA.

treated (FOR SYNTHETIC DATA ONLY) a logical indicator marking if the pathway was selected for treatment / synthetic disregulation. Obviously, real pathway results would not include such a column.

pVal_CNV for copy number variation data, these are the p-values for each pathway

pVal_RNAseq for RNAseq expression data, these are the p-values for each pathway

pVal_Prot for protein expression data, these are the p-values for each pathway

Details

This data represents some results from a simulation study we conducted to show the utility of the MiniMax technique. The results in this file are from a low-signal scenario (for the treatment group, 20% of the centered and scaled genes in the treated gene sets / pathways have been "upregulated"--i.e. shocked--by +0.2 standard deviations).

For the MiniMax statistic to work, the platform-specific p-values must come from a statistical test with well-controlled Type-I error rates.